Replaced __import__ with imp.load_source
Using __import__ does a global import and modules are not namespaced.
As a result if a module with a conflicting name is installed on the
system that module may be returned instead of the expected plugin.

Replaced with imp.load_source(module_name, filepath) which specifies
the source file to load with a given name, ensuring that the module
loaded is the plugin expected.